added side nav and pages #127
