+++ src/App.js
@@ -13,6 +13,9 @@ import Loading from './pages/Loading';
 import Login from './pages/Login';
 import Portal from './pages/Portal';
 import Dashboard from './pages/Dashboard';
+import Electricity from './pages/Electricity';
+import Gas from './pages/Gas';
+import Water from './pages/Water';
 import Settings from './pages/Settings';
 import ProfileDashboard from './pages/ProfileDashboard';
 import Events from './pages/Events';
@@ -41,6 +44,9 @@ export default function App() {
 					<Routes>
 						<Route path="/login" element={<Login />} />
 						<Route exact path="/" element={data && data.isLoggedIn ? <Dashboard hasSolarInverter={data.hasSolarInverters} /> : <Portal hasSolarInverter={data && data.hasSolarInverters} />} />
+						<Route path="/electricity" element={<Electricity hasSolarInverter={data.hasSolarInverters} />} />
+						<Route path="/gas" element={<Gas />} />
+						<Route path="/water" element={<Water />} />
 						<Route path="/profile" element={<ProfileDashboard />} />
 						<Route path="/events" element={<Events />} />
 						<Route path="settings" element={<Settings />} />
