+++ src/atoms/DashboardIcon/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function DashboardIcon() {
+	return (
+		<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 36 36">
+			<path fill="currentColor" d="M19.5,13.5V0H36v13.5H19.5z M0,19.5V0h16.5v19.5H0z M19.5,36V16.5H36V36H19.5z M0,36V22.5h16.5V36H0z M3,16.5h10.5V3H3V16.5z M22.5,33H33V19.5H22.5V33z M22.5,10.5H33V3H22.5V10.5z M3,33h10.5v-7.5H3V33z"/>
+		</svg>
+	);
+}
