+++ src/atoms/MenuIcon/index.js
@@ -1,0 +1,10 @@
+
+import React from 'react';
+
+export default function MenuIcon() {
+	return (
+		<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 36 24">
+			<path fill="currentColor" d="M0,24v-3h36v3H0z M0,13.5v-3h36v3H0z M0,3V0h36v3H0z"/>
+		</svg>
+	);
+}
