+++ src/pages/Dashboard/index.js
@@ -1,7 +1,5 @@
 import './dashboard.css';
-import React, { useState, useCallback } from 'react';
-
-import { DateTime } from 'luxon';
+import React, { useState } from 'react';
 
 import Default from '../../templates/Default';
 
@@ -9,7 +7,6 @@ import WidgetGrid from '../../organisms/WidgetGrid';
 
 import Widget from '../../molecules/Widget';
 
-import BoltIcon from '../../atoms/BoltIcon';
 import BoltArrowUpIcon from '../../atoms/BoltArrowUpIcon';
 import PowerPlugIcon from '../../atoms/PowerPlugIcon';
 import SunIcon from '../../atoms/SunIcon';
@@ -27,24 +24,6 @@ import ElectricityUsed from '../../molecules/ElectricityUsed';
 import CurrentSolarPowerGenerated from '../../molecules/SolarPowerGenerated';
 import GasUsage from '../../molecules/GasUsage';
 import WaterUsage from '../../molecules/WaterUsage';
-import ElectricityChart from '../../molecules/ElectricityChart';
-import GasChart from '../../molecules/GasChart';
-import CumulativeWaterUsageChart from '../../molecules/CumulativeWaterUsageChart';
-import WaterReceivedWeekChart from '../../molecules/WaterReceivedWeekChart';
-
-import Button from '../../atoms/Button';
-
-const getStartOfToday = () => {
-	const today = new Date();
-	today.setHours(0, 0, 0, 0);
-	return Math.round(today.getTime() / 1000);
-};
-
-const getEndOfToday = () => {
-	const today = new Date();
-	today.setHours(23, 59, 59, 999);
-	return Math.round(today.getTime() / 1000);
-};
 
 export default function Dashboard({ hasSolarInverter = false }) {
 	const [updatedCurrentElectricityReceived, setUpdatedCurrentElectricityReceived] = useState(null);
@@ -52,53 +31,6 @@ export default function Dashboard({ hasSolarInverter = false }) {
 	const [updatedCurrentSolarPowerGenerated, setUpdatedCurrentSolarPowerGenerated] = useState(null);
 	const [updatedCurrentElectricityUsing, setUpdatedCurrentElectricityUsing] = useState(null);
 	// TODO: weird solution, DateTime in props causes rerenders
-	const [electricityEightHourStartEnd] = useState({
-		start: Math.floor(DateTime.now().minus({ hours: 8 }).toSeconds()),
-		end: Math.floor(DateTime.now().toSeconds())
-	});
-	const [gasFourDaysStartEnd] = useState({
-		start: Math.floor(DateTime.now().minus({ days: 4 }).toSeconds()),
-		end: Math.floor(DateTime.now().toSeconds())
-	});
-	const [weekStartEnd] = useState({
-		start: Math.floor(DateTime.now().startOf('week').toSeconds()),
-		end: Math.floor(DateTime.now().endOf('week').toSeconds())
-	});
-	const [twelveMonthStartEnd] = useState({
-		start: Math.floor(DateTime.now().minus({ month: 11 }).startOf('month').toSeconds()),
-		end: Math.floor(DateTime.now().endOf('month').toSeconds())
-	});
-	const [fourYearsStartEnd] = useState({
-		start: Math.floor(DateTime.now().minus({ years: 4 }).toSeconds()),
-		end: Math.floor(DateTime.now().toSeconds())
-	});
-
-	const [waterChartDay, setWaterChartDay] = useState({
-		start: getStartOfToday(),
-		end: getEndOfToday()
-	});
-
-	const setMinus2 = useCallback(() => {
-		setWaterChartDay({
-			start: getStartOfToday() - (86400 * 2),
-			end: getEndOfToday() - (86400 * 2)
-		});
-	}, [setWaterChartDay]);
-
-	const setYesterday = useCallback(() => {
-		setWaterChartDay({
-			start: getStartOfToday() - 86400,
-			end: getEndOfToday() - 86400
-		});
-	}, [setWaterChartDay]);
-
-	const setToday = useCallback(() => {
-		setWaterChartDay({
-			start: getStartOfToday(),
-			end: getEndOfToday()
-		});
-	}, [setWaterChartDay]);
-
 	const updatedAtCurrentElectricityReceived = (timestamp) =>{
 		setUpdatedCurrentElectricityReceived(timestamp);
 	};
@@ -116,7 +48,7 @@ export default function Dashboard({ hasSolarInverter = false }) {
 	};
 
 	return (
-		<Default>
+		<Default title="Dashboard">
 			<WidgetGrid>
 				<Widget title="Receiving" name="current-electricity-received" updatedAt={updatedCurrentElectricityReceived} icon={<BoltArrowDownIcon />}>
 					<CurrentElectricityReceived updatedAt={updatedAtCurrentElectricityReceived} />
@@ -147,99 +79,6 @@ export default function Dashboard({ hasSolarInverter = false }) {
 				</Widget>
 				<Widget title="Today" name="water-usage" icon={<WaterIcon />}>
 					<WaterUsage />
-				</Widget>
-				<Widget title="Last 8 hours" name="electricity-usage-chart" icon={<BoltIcon />}>
-					<ElectricityChart
-						resolution="FIVE_MINUTES"
-						start={electricityEightHourStartEnd.start}
-						end={electricityEightHourStartEnd.end}
-						unit="WATT_HOUR"
-						softMax={100}
-						includeSolarPower={hasSolarInverter}
-					/>
-				</Widget>
-				<Widget title="Last 4 days" name="gas-usage-chart" icon={<FireIcon />}>
-					<GasChart
-						resolution="TWO_HOURS"
-						start={gasFourDaysStartEnd.start}
-						end={gasFourDaysStartEnd.end}
-						chartType="column"
-					/>
-				</Widget>
-				<Widget title="Today" name="cumulative-water-usage-chart" icon={<WaterIcon />}>
-					<ul style={{ display: 'flex', gap: '10px', listStyleType: 'none', padding: '0' }}>
-						<li><Button onClick={setMinus2} type="primary">-2</Button></li>
-						<li><Button onClick={setYesterday} type="primary">Yesterday</Button></li>
-						<li><Button onClick={setToday} type="primary">Today</Button></li>
-					</ul>
-					<CumulativeWaterUsageChart {...waterChartDay} />
-				</Widget>
-				<Widget title="Week" name="electrical-usage-current-week" icon={<BoltIcon />}>
-					<ElectricityChart
-						resolution="DAY"
-						start={weekStartEnd.start}
-						end={weekStartEnd.end}
-						timeFormat={{ weekday: 'long' }}
-						chartType="column"
-						includePrevious
-						softMax={2}
-					/>
-				</Widget>
-				<Widget title="Week" name="gas-usage-current-week" icon={<FireIcon />}>
-					<GasChart
-						title="Week"
-						resolution="DAY"
-						start={weekStartEnd.start}
-						end={weekStartEnd.end}
-						timeFormat={{ weekday: 'long' }}
-						chartType="column"
-						includePrevious
-						softMax={1}
-					/>
-				</Widget>
-				<Widget title="Week" name="water-usage-current-week" icon={<WaterIcon />}>
-					<WaterReceivedWeekChart />
-				</Widget>
-				<Widget title="Month" name="electrical-usage-by-month" icon={<BoltIcon />}>
-					<ElectricityChart
-						resolution="MONTH"
-						start={twelveMonthStartEnd.start}
-						end={twelveMonthStartEnd.end}
-						timeFormat={{ month: 'long' }}
-						chartType="column"
-						softMax={200}
-					/>
-				</Widget>
-				<Widget title="Month" name="gas-usage-by-month" icon={<FireIcon />}>
-					<GasChart
-						resolution="MONTH"
-						start={twelveMonthStartEnd.start}
-						end={twelveMonthStartEnd.end}
-						chartType="column"
-						timeFormat={{ month: 'long' }}
-						includePrevious
-						softMax={20}
-					/>
-				</Widget>
-				<Widget title="Year" name="electrical-usage-by-year" icon={<BoltIcon />}>
-					<ElectricityChart
-						resolution="YEAR"
-						start={fourYearsStartEnd.start}
-						end={fourYearsStartEnd.end}
-						timeFormat={{ year: 'numeric' }}
-						chartType="column"
-						softMax={1500}
-					/>
-				</Widget>
-				<Widget title="Year" name="gas-usage-by-year" icon={<FireIcon />}>
-					<GasChart
-						resolution="YEAR"
-						start={fourYearsStartEnd.start}
-						end={fourYearsStartEnd.end}
-						timeFormat={{ year: 'numeric' }}
-						chartType="column"
-						softMax={500}
-					/>
 				</Widget>
 			</WidgetGrid>
 		</Default>
