+++ src/templates/Settings/index.js
@@ -7,7 +7,7 @@ import SettingsHeader from '../../molecules/SettingsHeader';
 export default function Settings({ children, title = '', backButton = true }) {
 	return (
 		<div className={styles.default}>
-			<Header />
+			<Header title="Home Care" />
 			<SettingsHeader title={title} backButton={backButton} />
 			<main className={styles.pageContainer}>
 				{children}
